+++ libs/core/genblaze_core/__init__.py
@@ -96,6 +96,8 @@ _LAZY_IMPORTS: dict[str, tuple[str, str]] = {
     "RETRYABLE_STORAGE_CODES": ("genblaze_core.storage.errors", "RETRYABLE_STORAGE_CODES"),
     "classify_botocore_error": ("genblaze_core.storage.errors", "classify_botocore_error"),
     "KeyBuilder": ("genblaze_core.storage.key_builder", "KeyBuilder"),
+    "URLPolicy": ("genblaze_core.storage.url_policy", "URLPolicy"),
+    "URLPolicyError": ("genblaze_core.storage.url_policy", "URLPolicyError"),
     "ObjectMetadata": ("genblaze_core.storage.types", "ObjectMetadata"),
     "FileEntry": ("genblaze_core.storage.types", "FileEntry"),
     "ListPage": ("genblaze_core.storage.types", "ListPage"),